refactor: #145 extract entity cache from workspace into separate class
