+++ packages/core/src/lib/entity/workspace.ts
@@ -1,46 +1,24 @@
-import { cloneJson, permutateEntries } from "@entity-space/utils";
 import { Observable, Subject } from "rxjs";
-import { NamedCriteria } from "../criteria/public";
 import { mergeQueries, Query, reduceQueries } from "../query/public";
 import { IEntitySchema } from "../schema/schema.interface";
-import { createCriteriaTemplateForIndex } from "./create-criteria-template-for-index.fn";
 import { Entity } from "./entity";
+import { EntityCache } from "./entity-cache";
 import { IEntitySource } from "./entity-source.interface";
-import { EntityStore } from "./entity-store";
-import { expandEntities } from "./expand-entities.fn";
-import { flattenNamedCriteria } from "./flatten-named-criteria.fn";
-import { namedCriteriaToKeyPaths } from "./named-criteria-to-key-path.fn";
-import { normalizeEntities } from "./normalize-entities.fn";
 import { QueriedEntities } from "./queried-entities";
 
 // [todo] should implement IEntitySource
-export class Workspace {
+export class Workspace implements IEntitySource {
     private source?: IEntitySource;
-    private readonly stores = new Map<string, EntityStore>();
     private readonly queryCaches = new Map<string, Query[]>();
     private readonly queryCacheChanged = new Subject<Query[]>();
+    private readonly entityCache = new EntityCache();
 
     onQueryCacheChanged(): Observable<Query[]> {
         return this.queryCacheChanged.asObservable();
     }
 
-    addEntities(schema: IEntitySchema, entities: Entity[]): void {
-        entities = cloneJson(entities);
-        const normalized = normalizeEntities(schema, entities);
-
-        for (const schema of normalized.getSchemas()) {
-            const normalizedEntities = normalized.get(schema);
-            this.getOrCreateStore(schema).add(normalizedEntities);
-
-            // [todo] can not use until we implemented invert() @ named-criteria
-            // if (normalizedEntities.length > 0) {
-            //     const indexQueries = createQueriesFromEntities(schema, normalizedEntities);
-
-            //     for (const indexQuery of indexQueries) {
-            //         this.addExecutedQuery(indexQuery);
-            //     }
-            // }
-        }
+    add(schema: IEntitySchema, entities: Entity[]): void {
+        this.entityCache.addEntities(schema, entities);
     }
 
     setSource(source: IEntitySource): void {
@@ -79,7 +57,7 @@ export class Workspace {
         }
 
         if (entities.length > 0) {
-            this.addEntities(query.entitySchema, entities);
+            this.add(query.entitySchema, entities);
         }
     }
 
@@ -97,94 +75,30 @@ export class Workspace {
         return result;
     }
 
-    // [todo] remove any
-    async query(query: Query): Promise<false | Entity[]> {
+    async query(query: Query): Promise<false | QueriedEntities> {
         await this.loadUncachedIntoCache(query);
+        const entities = await this.queryAgainstCache(query);
 
-        return this.queryAgainstCache(query);
+        return new QueriedEntities(query, entities);
     }
 
     // [todo] remove any
     // [todo] should stay async because at one point i want to make use of service-workers
+    // [todo] should not exist at all? (or be private)
     async queryAgainstCache(query: Query): Promise<any[]> {
-        const schema = query.entitySchema;
-        const indexes = schema
-            .getIndexesIncludingKey()
-            .slice()
-            .sort((a, b) => b.getPath().length - a.getPath().length);
+        const result = await this.entityCache.query(query);
 
-        const criteriaTemplates = indexes.map(index => createCriteriaTemplateForIndex(index));
-        const [remappedCriteria] = query.criteria.remap(criteriaTemplates);
-
-        // [todo] remove "any" - but will result in compile error @ 02-loading-data.spec.ts
-        let entities: any[] = [];
-        const store = this.getOrCreateStore(schema);
-
-        if (remappedCriteria === false) {
-            entities = store.getAll();
-        } else {
-            for (const remappedCriterion of remappedCriteria) {
-                // [todo] we probably need to check for duplicates?
-                entities.push(...this.readFromStoreUsingIndexCriteria(store, remappedCriterion));
-            }
+        if (result === false) {
+            return [];
         }
 
-        if (Object.keys(query.expansion).length > 0) {
-            entities = cloneJson(entities); // [todo] dirty to do it here?
-
-            const self = this;
-
-            // [todo] hackity-hack. need instead an entity-source that always only goes against cache i guess?
-            // that would make workspace class a lot smaller maybe.
-            const expansionResult = await expandEntities(schema, query.expansion, entities, {
-                async query(q: Query) {
-                    const result = await self.queryAgainstCache(q);
-
-                    return new QueriedEntities(q, result);
-                },
-            });
-
-            console.log("[expansion-result]", expansionResult);
-        }
-
-        return query.criteria.filter(entities);
+        return result.getEntities();
     }
 
-    private readFromStoreUsingIndexCriteria(store: EntityStore, indexCriteria: NamedCriteria): Entity[] {
-        const bagKeyPaths = namedCriteriaToKeyPaths(indexCriteria);
-        const index = store.getIndexMatchingKeyPaths(bagKeyPaths);
-        const bagWithPrimitives = flattenNamedCriteria(indexCriteria);
-        const permutatedBags = permutateEntries(bagWithPrimitives);
-        const indexValues: (number | string)[][] = [];
-
-        for (const permutatedBag of permutatedBags) {
-            const indexValue: (number | string)[] = [];
-
-            for (const key of index.getPath()) {
-                indexValue.push(permutatedBag[key]);
-            }
-
-            indexValues.push(indexValue);
-        }
-
-        return store.getByIndexOrKey(index.getName(), indexValues);
-    }
-
-    clearCache(): void {
+    clear(): void {
         this.queryCaches.clear();
-        this.stores.clear();
+        this.entityCache.clear();
         this.queryCacheChanged.next([]);
-    }
-
-    private getOrCreateStore(schema: IEntitySchema): EntityStore {
-        let store = this.stores.get(schema.getId());
-
-        if (store === void 0) {
-            store = new EntityStore(schema);
-            this.stores.set(schema.getId(), store);
-        }
-
-        return store;
     }
 
     private getOrCreateQueryCache(schema: IEntitySchema): Query[] {
